+++ src/Symfony/Component/Form/Console/Helper/DescriptorHelper.php
@@ -12,9 +12,10 @@
 namespace Symfony\Component\Form\Console\Helper;
 
 use Symfony\Component\Console\Helper\DescriptorHelper as BaseDescriptorHelper;
+use Symfony\Component\ErrorHandler\ErrorRenderer\FileLinkFormatter;
 use Symfony\Component\Form\Console\Descriptor\JsonDescriptor;
 use Symfony\Component\Form\Console\Descriptor\TextDescriptor;
-use Symfony\Component\HttpKernel\Debug\FileLinkFormatter;
+use Symfony\Component\HttpKernel\Debug\FileLinkFormatter as LegacyFileLinkFormatter;
 
 /**
  * @author Yonel Ceruto <yonelceruto@gmail.com>
@@ -23,7 +24,7 @@ use Symfony\Component\HttpKernel\Debug\FileLinkFormatter;
  */
 class DescriptorHelper extends BaseDescriptorHelper
 {
-    public function __construct(FileLinkFormatter $fileLinkFormatter = null)
+    public function __construct(FileLinkFormatter|LegacyFileLinkFormatter $fileLinkFormatter = null)
     {
         $this
             ->register('txt', new TextDescriptor($fileLinkFormatter))